refactor: move getPartials to generate-html
